+++ packages/blocks/src/page-block/default/selection-manager/block-drag-handlers.ts
@@ -1,7 +1,7 @@
 import { assertExists } from '@blocksuite/global/utils';
 import type { PointerEventState } from '@blocksuite/lit';
 
-import type { DefaultSelectionManager } from './default-selection-manager.js';
+import type { DefaultSelectionManager } from './selection-manager.js';
 import { autoScroll } from './utils.js';
 
 export const BlockDragHandlers = {
