+++ packages/blocks/src/page-block/default/selection-manager/preview-drag-handlers.ts
@@ -2,7 +2,7 @@ import { assertExists } from '@blocksuite/global/utils';
 import type { PointerEventState } from '@blocksuite/lit';
 
 import { createDragEvent, Point } from '../../../__internal__/index.js';
-import type { DefaultSelectionManager } from './default-selection-manager.js';
+import type { DefaultSelectionManager } from './selection-manager.js';
 import { autoScroll } from './utils.js';
 
 export const PreviewDragHandlers = {
